Fix decoding of ResponseMessage
`result = Just Nothing` gets encoded to `result = null` in
JSON. However, upon decoding that gives us `result = Nothing` so we
cannot roundtrip. This is obviously undesirable and also breaks the
assumption that exactly one of `result` or `error` must be `Just`. As
far as I know, there is know way to get this behavior by setting Aeson
options, so I switched to a hand-written FromJSON instance for
`ResponseMessage`.