Use attoparsec to parse message headers incrementally
Previously, we were repeatedly reading one character, appending it to
the end and trying to parse until parsec returned a result. This is
needlessly wasteful and while it’s probably not a bottleneck, I think
the incremental parsing with attoparsec is also easier to understand
so it’s a win both in readability and performance.