Ensure that ResponseMessage has either a result or an error
Without this, it can happen that a request is parsed as a response
with both _result and _error set to Nothing. This has caused issues
for me with lsp-test which listens for a response message with the
same id after sending a request but actually ended up interpreting a
request send from the server as a response.