Tidy up some stuff in Core, add getVersionedTextDoc
When wanting to create textdocumentedits, you need to pass a versioned
text doc identifier. But often times the request that you are working
with only gives you an unversioned text doc identifier. This is a helper
function to convert it to the appropriate version by looking it up in
the VFS.