Capture: Eliminate racy recording of event captures
Previously we would use BSL.hPutStrLn without locking the handle. This
meant that parts of different events would become interleaved in the
log. We now use a TChan and dedicated writer thread for serialisation.

This also has the pleasant advantage of taking serialisation and writing
out of the latency-critical request-reply path.