Avoid unnecessary normalisation in Binary NormalizedFilePath
imho the Binary instance should be able to assume that the encoded
file path was already normalised which means that normalising again is
just unnecessarily slow.

Furthermore, it means that Binary instances
in ghcide fail to roundtrip since we rely on empty file paths. We
could fix that on the ghcide side but it’s somewhat brittle and I
don’t see a reason to not address this here.