Remove pointless IO from VFS
Setting up the VFS required IO just so it could capture a reference to a
temporary directory that could be used if we wanted to write files out.
This is weird and non-compositional: we can just ask the caller where
they want the files, and they can make the temporary directory if they
want.