Various fixes to progress handling (#539)
* Various fixes to progress handling

1. Server-initiated progress should wait for the client to acknowledge

This is an old bug. Per the spec, we're not allowed to send any reports
using the token if the client doesn't respond with a non-error response
to our creation of the token.

This is a bit subtle, because it means we may need to delay the sending
of the "begin" notification until we have received the token from the
client.

2. No easy way to use client-initiated progress

This is simpler and faster than server-initiated progress since you
don't need the extra message round-trip. You just need to pull out
the progress token (if there is one) from the request and use that.

I did two things to make this better:
- The progress functions now take the client token if there is one.
  If there isn't one they still fall back to server-initiated progress.
- The server capabilities can now advertise that client-initiated
  progress is supported.

* Make the default for client-initiated progress to be off

* Don't lose progress updates if it takes a while to start

* Refactor tests

* Add test for cancellation

* Fix minor bug in lsp-test config handling

* More tests

* Try this

* Fix formatting

* Add comment and purge traces