Accept null in place of a missing value
We sometimes see reports where HLS is not working because the client
sends us slightly wrong JSON. We should follow the principle of
robustness and try to accept this where we can - following the spec
precisely is hard!

The specific case we hadle here is accepting null to mean "missing"
when null is not a valid value for the type. Currently we will treat
it as a _present_ value and then fail to interpret null as a value of
the type.