Always try to register for didChangeConfiguration  (#548)
* Register for workspace/didChangeConfiguration always

The new direction of the spec seems to be that you should always do
this. In any case, it doesn't hurt.

`lsp-test` has some changes to adapt to the increased registration
message spam, and also changes to be a more picky client, refusing to
send notifications unless you _do_ register. This is good, since it
forces us to handle the most annoying version of client behaviour.

* Add bound on extra

* fix