Reject messages that come in after 'shutdown'
This is mandated by the spec so we can do it. We also expose the
shutdown barrier, which I think can be convenient. For more
sophisticated usecases people should just install a proper shutdown
handler.

I tried to write some tests for this, but I gave up because it requires
some singificant surgery on `lsp-test`, which stops recording messages
when it receives `shutdown` :(