Rewrite progress handling to allow for debouncing messages (#571)
* Rewrite progress handling to allow for debouncing messages

This had to be redone in order to allow us to "wake up" and notice that
there are pending messages. I also wrote it so there can be a stateful
interface (the `ProgressTracker`) which I think might make it easier to
use in that weird case in `ghcide`. I haven't exposed that yet, though.

* Remove stateful interface

* Delay sending the create request also

* Changelog

* Move progress code to its own module